requirements.txt
streamlit
openai-agents
python-dotenv
qdrant-client
agno
asyncio
pypdf
yagmail
redis 
redisvl
pytesseract 
pdf2image 
PyPDF2
